apps/mcp-servers/market-data/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finsight-mcp-market-data"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",
  "fastmcp>=0.4",
  "openbb>=4.0",
  "openbb-yfinance",
  "httpx>=0.27",
  "redis[hiredis]>=5.0",
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "structlog>=24.1",
  "fakeredis>=2.23",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
